+++ src/internal/admin.py
@@ -3,6 +3,7 @@ from django.db.models.functions import Concat
 from django.utils.translation import gettext_lazy as _
 from simple_history.admin import SimpleHistoryAdmin
 
+from util import html_utils
 from util.admin_utils import DefaultAdminWidgetsMixin, UserSearchFieldsMixin, search_escaped_and_unescaped
 from .models import Member, Quote, Secret, SystemAccess
 
@@ -36,11 +37,38 @@ class SystemAccessAdmin(UserSearchFieldsMixin, admin.ModelAdmin):
 
 
 class SecretAdmin(DefaultAdminWidgetsMixin, SimpleHistoryAdmin):
-    list_display = ('title', 'priority', 'last_modified')
+    list_display = ('title', 'priority', 'get_extra_view_permissions', 'last_modified')
+    list_filter = (
+        ('extra_view_permissions', admin.RelatedOnlyFieldListFilter),
+    )
     search_fields = ('title', 'content')
     list_editable = ('priority',)
 
+    filter_horizontal = ('extra_view_permissions',)
     readonly_fields = ('last_modified',)
+
+    @admin.display(description=_("extra view permissions"))
+    def get_extra_view_permissions(self, secret: Secret):
+        return html_utils.block_join(secret.extra_view_permissions.all(), sep="<b>&bull;</b>") or None
+
+    def has_add_permission(self, request):
+        return self._has_permission(super().has_add_permission(request), request)
+
+    def has_change_permission(self, request, obj: Secret = None):
+        return self._has_permission(super().has_change_permission(request, obj), request, obj)
+
+    def has_delete_permission(self, request, obj: Secret = None):
+        return self._has_permission(super().has_delete_permission(request, obj), request, obj)
+
+    def has_view_permission(self, request, obj: Secret = None):
+        return self._has_permission(super().has_view_permission(request, obj), request, obj)
+
+    @staticmethod
+    def _has_permission(has_base_perm, request, obj: Secret = None):
+        if not obj:
+            return has_base_perm
+
+        return has_base_perm and request.user.has_perms(obj.extra_view_perms_str_tuple)
 
     def get_queryset(self, request):
         return super().get_queryset(request).default_order_by()