+++ src/internal/tests/test_views.py
@@ -2,6 +2,9 @@ from http import HTTPStatus
 from urllib.parse import urlparse
 
 from django.conf import settings
+from django.contrib.auth.models import Permission
+from django.contrib.contenttypes.models import ContentType
+from django.db import IntegrityError, transaction
 from django.templatetags.static import static
 from django.test import Client, TestCase, override_settings
 from django.urls import clear_url_caches
@@ -9,10 +12,12 @@ from django.urls import clear_url_caches
 from contentbox.forms import EditSourceContentBoxForm
 from contentbox.models import ContentBox
 from users.models import User
-from util.auth_utils import get_perms
+from util.auth_utils import get_perms, perm_to_str
 from util.url_utils import reverse_internal
 from web.multilingual.data_structures import MultiLingualTextStructure
+from web.tests.test_urls import ADMIN_CLIENT_DEFAULTS, reverse_admin
 from .test_urls import INTERNAL_CLIENT_DEFAULTS
+from ..models import Secret
 
 
 class InternalContentBoxTests(TestCase):
@@ -108,3 +113,109 @@ class InternalContentBoxTests(TestCase):
                 self.assertEqual(len(content_text_structure.languages), len(expected_content_languages))
                 for language in expected_content_languages:
                     self.assertEqual(content_text_structure[language], expected_content_languages[language])
+
+
+class SecretTests(TestCase):
+
+    def setUp(self):
+        self.normal_secret = Secret.objects.create(title="Storage key code", content="1234")
+        self.board_secret = Secret.objects.create(title="Code for deleting the organization", content="asdf")
+
+        self.board_perm = Permission.objects.create(
+            codename='board_perm',
+            name="Can view secrets that only members of the board need",
+            content_type=ContentType.objects.get_for_model(Secret),
+        )
+        self.board_secret.extra_view_permissions.add(self.board_perm)
+
+        self.member_user = User.objects.create_user("member_user")
+        self.board_member_user = User.objects.create_user("board_member_user")
+        # This superuser should not have any permissions added to their `user_permissions` or through groups,
+        # to test that they have all permissions without having to grant any explicitly
+        self.superuser = User.objects.create_user("superuser", is_staff=True, is_superuser=True)
+
+        base_secret_perms = ('internal.is_internal', 'internal.view_secret', 'internal.change_secret', 'internal.delete_secret')
+        self.member_user.add_perms(*base_secret_perms)
+        self.board_member_user.add_perms(*base_secret_perms, perm_to_str(self.board_perm))
+
+        self.member_client = Client(**INTERNAL_CLIENT_DEFAULTS)
+        self.board_member_client = Client(**INTERNAL_CLIENT_DEFAULTS)
+        self.superuser_client = Client(**INTERNAL_CLIENT_DEFAULTS)
+
+        self.member_client.force_login(self.member_user)
+        self.board_member_client.force_login(self.board_member_user)
+        self.superuser_client.force_login(self.board_member_user)
+
+    def test_only_secrets_users_have_permission_to_view_are_listed(self):
+        secret_list_url = reverse_internal('secret_list')
+
+        def assert_only_secrets_user_has_permission_to_view_are_listed(client: Client, *, board_member: bool):
+            response = client.get(secret_list_url)
+            self.assertContains(response, self.normal_secret.title)
+            self.assertContains(response, self.normal_secret.content)
+            # `count=None` means at least once
+            self.assertContains(response, self.board_secret.title, count=None if board_member else 0)
+            self.assertContains(response, self.board_secret.content, count=None if board_member else 0)
+
+        assert_only_secrets_user_has_permission_to_view_are_listed(self.member_client, board_member=False)
+        assert_only_secrets_user_has_permission_to_view_are_listed(self.board_member_client, board_member=True)
+        assert_only_secrets_user_has_permission_to_view_are_listed(self.superuser_client, board_member=True)
+
+    def test_users_can_only_change_secrets_they_have_permission_to_view(self):
+        def assert_users_can_change_secret(secret: Secret, *, is_board_secret: bool):
+            change_url = reverse_internal('edit_secret', secret.pk)
+            self.assertEqual(self.member_client.get(change_url).status_code, HTTPStatus.FORBIDDEN if is_board_secret else HTTPStatus.OK)
+            self.assertEqual(self.board_member_client.get(change_url).status_code, HTTPStatus.OK)
+            self.assertEqual(self.superuser_client.get(change_url).status_code, HTTPStatus.OK)
+
+        assert_users_can_change_secret(self.normal_secret, is_board_secret=False)
+        assert_users_can_change_secret(self.board_secret, is_board_secret=True)
+
+    def test_users_can_only_delete_secrets_they_have_permission_to_view(self):
+        def assert_user_can_delete_secret(client: Client, secret: Secret, *, can_delete: bool):
+            delete_url = reverse_internal('delete_secret', secret.pk)
+            try:
+                with transaction.atomic():
+                    # `FOUND` means that the request was successful and that the client is redirected to the view's `success_url`
+                    self.assertEqual(client.delete(delete_url).status_code, HTTPStatus.FOUND if can_delete else HTTPStatus.FORBIDDEN)
+                    self.assertEqual(Secret.objects.filter(pk=secret.pk).exists(), not can_delete)
+                    # Raise an error so that the transaction is rolled back
+                    raise IntegrityError
+            except IntegrityError:
+                pass
+            # Ensure that the deletion of the secret was rolled back (not strictly necessary to test)
+            self.assertTrue(Secret.objects.filter(pk=secret.pk).exists())
+
+        assert_user_can_delete_secret(self.member_client, self.normal_secret, can_delete=True)
+        assert_user_can_delete_secret(self.board_member_client, self.normal_secret, can_delete=True)
+        assert_user_can_delete_secret(self.superuser_client, self.normal_secret, can_delete=True)
+
+        assert_user_can_delete_secret(self.member_client, self.board_secret, can_delete=False)
+        assert_user_can_delete_secret(self.board_member_client, self.board_secret, can_delete=True)
+        assert_user_can_delete_secret(self.superuser_client, self.board_secret, can_delete=True)
+
+    def test_staff_can_only_view_or_delete_secrets_they_have_permission_to_view(self):
+        # Make the users staff
+        self.member_user.is_staff = True
+        self.member_user.save()
+        self.board_member_user.is_staff = True
+        self.board_member_user.save()
+
+        member__admin_client = Client(**ADMIN_CLIENT_DEFAULTS)
+        board_member__admin_client = Client(**ADMIN_CLIENT_DEFAULTS)
+        superuser__admin_client = Client(**ADMIN_CLIENT_DEFAULTS)
+
+        member__admin_client.force_login(self.member_user)
+        board_member__admin_client.force_login(self.board_member_user)
+        superuser__admin_client.force_login(self.superuser)
+
+        def assert_users_can_view_admin_page(path: str, *, is_board_secret: bool):
+            self.assertEqual(member__admin_client.get(path).status_code, HTTPStatus.FORBIDDEN if is_board_secret else HTTPStatus.OK)
+            self.assertEqual(board_member__admin_client.get(path).status_code, HTTPStatus.OK)
+            self.assertEqual(superuser__admin_client.get(path).status_code, HTTPStatus.OK)
+
+        assert_users_can_view_admin_page(reverse_admin('internal_secret_change', args=[self.normal_secret.pk]), is_board_secret=False)
+        assert_users_can_view_admin_page(reverse_admin('internal_secret_change', args=[self.board_secret.pk]), is_board_secret=True)
+
+        assert_users_can_view_admin_page(reverse_admin('internal_secret_delete', args=[self.normal_secret.pk]), is_board_secret=False)
+        assert_users_can_view_admin_page(reverse_admin('internal_secret_delete', args=[self.board_secret.pk]), is_board_secret=True)